requirements.txt
%%writefile requirements.txt
torch
torchvision
numpy
Pillow
opencv-python